feat: Begin implementing “MP as a kit” and next gen architecture
