+++ mParticle-Apple-SDK/MPBackendController.h
@@ -1,5 +1,6 @@
 #import "MPEnums.h"
 #import "MPNetworkCommunication.h"
+#import "MPEventLogging.h"
 
 #if TARGET_OS_IOS == 1
     #import "MPNotificationController.h"
@@ -30,22 +31,6 @@ typedef NS_ENUM(NSUInteger, MPProfileChange) {
     MPProfileChangeDelete
 };
 
-typedef NS_ENUM(NSUInteger, MPExecStatus) {
-    MPExecStatusSuccess = 0,
-    MPExecStatusFail,
-    MPExecStatusMissingParam,
-    MPExecStatusDisabledRemotely,
-    MPExecStatusEnabledRemotely,
-    MPExecStatusOptOut,
-    MPExecStatusDataBeingFetched,
-    MPExecStatusInvalidDataType,
-    MPExecStatusDataBeingUploaded,
-    MPExecStatusServerBusy,
-    MPExecStatusItemNotFound,
-    MPExecStatusDisabledInSettings,
-    MPExecStatusNoConnectivity
-};
-
 extern const NSTimeInterval kMPRemainingBackgroundTimeMinimumThreshold;
 extern const NSInteger kNilAttributeValue;
 extern const NSInteger kExceededAttributeCountLimit;
@@ -64,7 +49,6 @@ extern const NSInteger kInvalidKey;
 #endif
 
 @property (nonatomic, weak, nullable) id<MPBackendControllerDelegate> delegate;
-@property (nonatomic, strong, nullable) NSMutableSet<MPEvent *> *eventSet;
 @property (nonatomic, strong, nullable) MPNetworkCommunication *networkCommunication;
 @property (nonatomic, strong, nullable) MPSession *session;
 @property (nonatomic, readwrite) NSTimeInterval sessionTimeout;
@@ -79,18 +63,9 @@ extern const NSInteger kInvalidKey;
 - (nonnull MParticleSession *)tempSession;
 - (void)beginTimedEvent:(nonnull MPEvent *)event completionHandler:(void (^ _Nonnull)(MPEvent * _Nonnull event, MPExecStatus execStatus))completionHandler;
 + (BOOL)checkAttribute:(nonnull NSDictionary *)attributesDictionary key:(nonnull NSString *)key value:(nonnull id)value error:(out NSError *__autoreleasing _Nullable * _Nullable)error;
-- (nullable MPEvent *)eventWithName:(nonnull NSString *)eventName;
 - (nullable NSString *)execStatusDescription:(MPExecStatus)execStatus;
 - (nullable NSNumber *)incrementSessionAttribute:(nonnull MPSession *)session key:(nonnull NSString *)key byValue:(nonnull NSNumber *)value;
 - (nullable NSNumber *)incrementUserAttribute:(nonnull NSString *)key byValue:(nonnull NSNumber *)value;
-- (void)leaveBreadcrumb:(nonnull MPEvent *)event completionHandler:(void (^ _Nonnull)(MPEvent * _Nonnull event, MPExecStatus execStatus))completionHandler;
-- (void)logError:(nullable NSString *)message exception:(nullable NSException *)exception topmostContext:(nullable id)topmostContext eventInfo:(nullable NSDictionary *)eventInfo completionHandler:(void (^ _Nonnull)(NSString * _Nullable message, MPExecStatus execStatus))completionHandler;
-- (void)logCrash:(nullable NSString *)message stackTrace:(nullable NSString *)stackTrace plCrashReport:(nonnull NSString *)plCrashReport completionHandler:(void (^ _Nonnull)(NSString * _Nullable message, MPExecStatus execStatus)) completionHandler;
-- (void)logBaseEvent:(nonnull MPBaseEvent *)event completionHandler:(void (^ _Nonnull)(MPBaseEvent * _Nonnull event, MPExecStatus execStatus))completionHandler;
-- (void)logEvent:(nonnull MPEvent *)event completionHandler:(void (^ _Nonnull)(MPEvent * _Nonnull event, MPExecStatus execStatus))completionHandler;
-- (void)logCommerceEvent:(nonnull MPCommerceEvent *)commerceEvent completionHandler:(void (^ _Nonnull)(MPCommerceEvent * _Nonnull commerceEvent, MPExecStatus execStatus))completionHandler;
-- (void)logNetworkPerformanceMeasurement:(nonnull MPNetworkPerformance *)networkPerformance completionHandler:(void (^ _Nullable)(MPNetworkPerformance * _Nonnull networkPerformance, MPExecStatus execStatus))completionHandler;
-- (void)logScreen:(nonnull MPEvent *)event completionHandler:(void (^ _Nonnull)(MPEvent * _Nonnull event, MPExecStatus execStatus))completionHandler;
 - (void)setOptOut:(BOOL)optOutStatus completionHandler:(void (^ _Nonnull)(BOOL optOut, MPExecStatus execStatus))completionHandler;
 - (MPExecStatus)setSessionAttribute:(nonnull MPSession *)session key:(nonnull NSString *)key value:(nonnull id)value;
 - (void)setUserTag:(nonnull NSString *)key timestamp:(nonnull NSDate *)timestamp completionHandler:(void (^ _Nullable)(NSString * _Nonnull key, id _Nullable value, MPExecStatus execStatus))completionHandler;
@@ -101,7 +76,6 @@ extern const NSInteger kInvalidKey;
 - (void)startWithKey:(nonnull NSString *)apiKey secret:(nonnull NSString *)secret firstRun:(BOOL)firstRun installationType:(MPInstallationType)installationType proxyAppDelegate:(BOOL)proxyAppDelegate startKitsAsync:(BOOL)startKitsAsync consentState:(MPConsentState *_Nullable)consentState completionHandler:(dispatch_block_t _Nonnull)completionHandler;
 - (void)unproxyOriginalAppDelegate;
 - (void)saveMessage:(nonnull MPMessage *)message updateSession:(BOOL)updateSession;
-- (void)skipNextUpload;
 - (MPExecStatus)waitForKitsAndUploadWithCompletionHandler:(void (^ _Nullable)(void))completionHandler;
 - (nonnull NSMutableDictionary<NSString *, id> *)userAttributesForUserId:(nonnull NSNumber *)userId;
 - (nonnull NSMutableArray<NSDictionary<NSString *, id> *> *)userIdentitiesForUserId:(nonnull NSNumber *)userId;
