+++ mParticle-Apple-SDK/Next Gen/MPComponent.h
@@ -1,0 +1,16 @@
+//
+//  MPComponent.h
+//  mParticle-Apple-SDK
+//
+//  Created by Ben Baron on 6/29/23.
+//
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@protocol MPComponent <NSObject>
+
+@end
+
+NS_ASSUME_NONNULL_END
